+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.1
 - Add `Chat` methods `read`, `add_operator`, `remove_operator`, `join`, `exit`, `invite`, and `kick`
+- Add `ChatUser` get method and getters
 
 ### 0.1.0
 - Add basic (mostly read-only) classes for objects